rebase master, utils
